+++ CHANGELOG.md
@@ -5,7 +5,7 @@
   - made tests comprehensive
   - optimized test duration by including both match by id and by name in the same test
   - removed duplicated code from import, data_source, resource and tests files
-- **new features**: import for `snapshot`, data_source for `group`, `user`
+- **new features**: import for `snapshot`, `ipblock`, data_source for `group`, `user`, `ipblock` 
 
 ## 6.0.0-beta.13
 - **bug fixes**: fixed issue #112 can't attach existing volume to server after recreating server